fix: normaliza tipos numpy para serialização JSON
- Converte explicitamente valores numpy (uint8, int32, float32, etc.) para tipos Python nativos
- Corrige erro de serialização TypeError: 'numpy.uint8' object is not iterable
- Atualiza serviços: colors, texture, scene, lighting, depth, saliency, emotion, cta
- Garante que todos os valores retornados são serializáveis pelo FastAPI
- Atualiza CHANGELOG.md com as correções

+++ services/cta.py
@@ -132,14 +132,14 @@ class CTAService:
             
             effectiveness_scores.append(score)
         
-        avg_score = np.mean(effectiveness_scores) if effectiveness_scores else 0.0
+        avg_score = float(np.mean(effectiveness_scores)) if effectiveness_scores else 0.0
         
         if avg_score < 0.5:
             recommendations.append("Melhore visibilidade do CTA aumentando contraste e tamanho")
         
         return {
             "cta_present": True,
-            "cta_count": len(cta_elements),
+            "cta_count": int(len(cta_elements)),
             "effectiveness_score": round(avg_score, 2),
             "recommendations": recommendations if recommendations else ["CTAs bem posicionados e visíveis"]
         }
